+++ Properties/PropertiesMethods.js
@@ -1,5 +1,7 @@
 module.exports = {
-    isInt: value => parseInt(value) !== NaN && parseInt(value) == value,
-    isString: value => typeof value === "string",
-    isBoolean: value => typeof value === "boolean"
+    isInt: v => parseInt(v) !== NaN && parseInt(v) == v,
+    isString: v => typeof v === "string",
+    isBoolean: v => typeof v === "boolean",
+    isFunction: v => typeof v === "function",
+    isDateTime: v => new Date(v).getDate() !== NaN
 }