+++ Properties/StringProperties.js
@@ -1,22 +1,32 @@
-const { isInt, isString, isBoolean } = require("./PropertiesMethods")
+const { isInt, isString, isBoolean, isFunction } = require("./PropertiesMethods")
 
 const Properties = {
   type: value => value === "string",
+  transform: isFunction,
   default: isString,
+  unique: isBoolean,
   required: isBoolean,
   minLength: isInt,
   maxLength: isInt,
+  uppercase: isBoolean,
+  lowercase: isBoolean,
 }
 
 const Validators = {
-  type: value => { state: isString(value) },
-  default: (value, property) => value == null || value === "" ? {
+  type: value => ({ state: isString(value) }),
+  transform: (value, property) => ({state: true, return: property(value)}),
+  default: (value, property) => (value == null || value === "" ? {
     state: true,
     return: property
-  } : { state: true },
-  required: value => { state: !(value === "" || !isString(value)) },
-  minLength: (value, property) => { state: value.length >= property },
-  maxLength: (value, property) => { state: value.length <= property },
+  } : { state: true }),
+  unique: (value, property, object, fieldName) => (property ? 
+    { state: object.findData({[fieldName]: value}).length === 0} 
+  : { state: true }),
+  required: (value, property) => (property ? { state: !(value === "" || !isString(value)) } : {state: true}),
+  minLength: (value, property) => ({ state: value.length >= property }),
+  maxLength: (value, property) => ({ state: value.length <= property }),
+  uppercase: (value, property) => ({ state: true, return: property ? value.toUpperCase() : value }),
+  lowercase: (value, property) => ({ state: true, return: property ? value.toLowerCase() : value })
 }
 
 module.exports = {