+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v0.2.0 (10.07.2021)
+
+### Features
+
+- Tables Main Methods (Create and Read (Find with Regex))
+- Provide unique and autoIncrement properties
+- Add specific support for dateTime
+
+### Improvements
+
+- Remove Required specific check
+- Checking with config to make all properties work
+
 ## v0.1.0 (06.07.2021)
 
 ### Features